+++ site/src/data/campusData.js
@@ -1,0 +1,38 @@
+export const campusData = [
+    {
+      id: 1,
+      name: 'Sup de Vinci',
+      occupationRate: 75,
+      totalRooms: 24,
+      availableRooms: ['salle 11', 'salle 15', 'salle 3', 'salle labo 1'],
+      occupiedRooms: 18,
+      color: '#4ECDC4'
+    },
+    {
+      id: 2,
+      name: 'ESILV Campus',
+      occupationRate: 60,
+      totalRooms: 20,
+      availableRooms: ['salle A12', 'salle B5', 'amphithéâtre 2'],
+      occupiedRooms: 12,
+      color: '#45B7D1'
+    },
+    {
+      id: 3,
+      name: 'EPITECH',
+      occupationRate: 85,
+      totalRooms: 18,
+      availableRooms: ['lab 3', 'room 201'],
+      occupiedRooms: 15,
+      color: '#96CEB4'
+    },
+    {
+      id: 4,
+      name: 'IIM Digital',
+      occupationRate: 40,
+      totalRooms: 15,
+      availableRooms: ['studio 1', 'studio 2', 'salle créa', 'lab mac'],
+      occupiedRooms: 6,
+      color: '#FFEAA7'
+    }
+  ];
