+++ site/src/data/mockData.js
@@ -1,0 +1,23 @@
+// mockData.js
+
+export const groupes = [
+  { id: 1, nom: 'A1 ESILV' },
+  { id: 2, nom: 'A2 ESILV' },
+  { id: 3, nom: 'A3 ESILV' },
+  { id: 4, nom: 'B1 EPITECH' },
+  { id: 5, nom: 'B2 EPITECH' },
+  { id: 6, nom: 'M1 IIM CDA' },
+  { id: 7, nom: 'M2 IIM CDA' }
+];
+
+export const salles = [
+  { id: 1, nom: 'Sup de vinci - 204' },
+  { id: 2, nom: 'Informatique - 154' },
+  { id: 3, nom: 'Outlook - 3' },
+  { id: 4, nom: 'Sup de vinci - 1' }
+];
+
+export const notifications = [
+  { id: 1, type: 'success', message: 'Sup de Vinci - 204 | 6 mai de 8h à 9h', detail: 'Votre réservation a été acceptée.' },
+  { id: 2, type: 'info', message: 'Sup de Vinci - 204 | 6 mai de 10h à 14h', detail: 'Vous avez été ajouté à un cours.' }
+];
